Adding date functions, and a failed attempt at using LinkedHashMap
The idea here is/was to use a LinkedHashMap as a way of sending chart
data to a function that will render the small multiples charts in as
generic a way as possible. I was originally thinking that a
LinkedHashMap would give me a data structure that contained the row
label String,  associated with an ArrayList of SnapEntrys in a defined
order. Problem is though that LinkedHashMaps don't give you an index
for each entry (hence the linked part?). Without the index I still need
to include or  get an indexable List (a la StringList) so why not just
send a StringList and a Hashmap both to the chart function and avoid
LinkedHashMap entirely?