In process to creating a generalized charts rendering function.
Its not pretty but it works. In the midst of creating a more generic
rendering function for the small multiples charts. Right now it needs
some manual tweaking to switch between rendering based on days of the
week vs a room list but it's on it's way to a generalized solution.

Also cleaned up some old debugging code that's been lingering for a
while and not being used.